Refactor DTOs and entities; rename fields in BillDTO, update quantity type in ProductsDTO, adjust @ToString annotations, and enhance SQL script with additional queries

+++ gadget-plus/src/main/java/com/debugeandoideas/gadgetplus/dto/ProductsDTO.java
@@ -5,11 +5,13 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigInteger;
+
 @Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class ProductsDTO {
-    private Long quantity;
+    private BigInteger quantity;
     private String name;
 }
